aquadem/requirements.txt
dm-acme==0.3.0
dm-control==0.0.425341097
dm-env==1.5
dm-haiku==0.0.5
dm-reverb==0.6.1
dm-sonnet==2.0.0
flax==0.4.0
gym==0.21.0
jax==0.2.28
numpy==1.21.5
optax==0.1.1
rlds==0.1.3
tensorflow==2.7.0
tensorflow_datasets==4.4.0
git+https://github.com/deepmind/rlax@master
git+https://github.com/rail-berkeley/d4rl@master
